requirements.txt
torch
torchvision
transformers
accelerate
einops
huggingface_hub
gradio
numpy
wandb
tqdm
beautifulsoup4==4.12.0
discord.py==2.2.2
evaluate==0.4.0
fastapi==0.98.0
langchain==0.0.154
nltk==3.8.1
nbconvert==7.6.0
nbformat==5.9.0
markdown==3.4.4
pandas==1.5.3
python-dotenv==1.0.0
Requests==2.29.0
scikit_learn==1.2.2
sentence-transformers==2.2.2
InstructorEmbedding==1.0.0
faiss_cpu==1.7.3
uvicorn==0.22.0
pytest==7.3.1
google-cloud-bigquery==3.17.2